Code to trim the scale and offset of the VDAC.
Ensure to change the device addresses and functions of the measuring instruments while using the code. Also ensure to change the file path where data is stored. This code just prints out the scale and offset trims. It does not write the values into the trim register or fuse the values into the device.